+++ go/backend/kv_file/kv_file_mocks.go
@@ -3,15 +3,16 @@
 //
 // Generated by this command:
 //
-//	mockgen -source kv_file.go -destination kv_file_mocks.go -package common
+//	mockgen -source kv_file.go -destination kv_file_mocks.go -package kv_file
 //
 
-// Package common is a generated GoMock package.
-package common
+// Package kv_file is a generated GoMock package.
+package kv_file
 
 import (
 	reflect "reflect"
 
+	common "github.com/0xsoniclabs/carmen/go/common"
 	gomock "go.uber.org/mock/gomock"
 )
 
@@ -253,10 +254,10 @@ func (mr *MockKVFileWithMemoryFootprintMockRecorder[K, V]) GetAll() *gomock.Call
 }
 
 // GetMemoryFootprint mocks base method.
-func (m *MockKVFileWithMemoryFootprint[K, V]) GetMemoryFootprint() *MemoryFootprint {
+func (m *MockKVFileWithMemoryFootprint[K, V]) GetMemoryFootprint() *common.MemoryFootprint {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "GetMemoryFootprint")
-	ret0, _ := ret[0].(*MemoryFootprint)
+	ret0, _ := ret[0].(*common.MemoryFootprint)
 	return ret0
 }
